+++ packages/components/bolt-svg-animations/src/svg/index.js
@@ -1,3 +1,5 @@
 export * from './automation';
 export * from './orbit';
 export * from './radar';
+export * from './connectionBand';
+export * from './tripleConnectionBand';