+++ packages/components/bolt-link/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.6.0](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-link/compare/v2.6.0-beta.2...v2.6.0) (2019-08-30)
+
+
+### Bug Fixes
+
+* prevent id attribute from being duplicated inside link component ([6a63a50](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-link/commit/6a63a50))
+
+
+
+
+
+# [2.6.0-beta.2](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-link/compare/v2.6.0-beta.1...v2.6.0-beta.2) (2019-08-27)
+
+**Note:** Version bump only for package @bolt/components-link
+
+
+
+
+
+# [2.6.0-beta.1](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-link/compare/v2.5.6...v2.6.0-beta.1) (2019-08-09)
+
+
+### Bug Fixes
+
+* correctly add all dependencies ([d9de9c0](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-link/commit/d9de9c0))
+* remove outline from button and link custom-elements to fix double-focus ([5454d1f](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-link/commit/5454d1f))
+
+
+
+
+
 ## [2.5.6](https://github.com/bolt-design-system/bolt/tree/master/packages/components/bolt-link/compare/v2.5.5...v2.5.6) (2019-07-30)
 
 **Note:** Version bump only for package @bolt/components-link